Can not create a scene with Physics enable.
The important thing to realise is that a scene by default is not created with physics enabled in it.
One has to create a scene for the static method ```Scene::createWithPhysics()``` . Otherwise, I noted that calling any physics methods on a Scene without physics support made the program exit with error code 11. Should find more on that sometime later.

But now the question that I need to answer is, what if I only want certain parts of the Scene to have physics. I guess I could add a physics enabled scene as a child to a normal scene. Scene dimensions???